PASS Test6-System Balance 5 w/1 Nickel Inserted
When a Coin of type NICKEL is inserted into a CoinController instance,
    the system balance will be set to five cents independent of the
    previous system balance.

CREATE int field in CoinController to hold the current system balance
    in cents.  This field is initialized to zero cents.

CREATE CoinController.insert() to process all Coin insertions.  This
    method accepts one Coin enum constant as an argument, and sets the
    current system balance to five cents.

CREATE Coin enum with NICKEL as the only declared constant.  A NICKEL
    can be inserted into CoinController.insert() to add value to the
    system balance.

MODIFY CoinController.getSystemBalanceInCents() to return the int field
    that holds the system balance in cents.