PASS Test8-System Balance 10 w/1 Dime Inserted
When a Coin of type DIME is inserted into a CoinController instance,
    ten cents will be added to the current system balance.

MODIFY CoinController.insert() to determine the inserted Coins' identities
    using their weights, diameters, and thicknesses.  If the Coin is a NICKEL,
    five cents will be added to the system balance.  Else, ten cents will be
    added to the system balance.

MODIFY Coin so that each declared constant is associated with three
    double fields:
    -Coin Weight in Grams
    -Coin Diameter in MM
    -Coin Thickness in MM

POPULATE Coin.NICKEL with values for weight, diameter, and thickness.

CREATE Coin enum constant DIME, and populate with dummy zero values.